feat: refactor, rename to Admin
