+++ src/Controller/Admin/AsyncJobsController.php
@@ -1,5 +1,5 @@
 <?php
-namespace App\Controller\Administration;
+namespace App\Controller\Admin;
 
 /**
  * Async Jobs Controller
