+++ src/Controller/Admin/ConfigController.php
@@ -1,5 +1,5 @@
 <?php
-namespace App\Controller\Administration;
+namespace App\Controller\Admin;
 
 /**
  * Config Controller
