+++ src/Template/Element/Admin/index_content.twig
@@ -1,4 +1,4 @@
-{% element ('Administration/sidebar') %}
+{% element ('Admin/sidebar') %}
 
 {% do _view.assign('title', 'Administration ' ~ resourceType|humanize) %}
 {% do _view.assign('bodyViewClass',  'view-module view-model') %}
