+++ src/Template/Pages/Admin/Endpoints/index.twig
@@ -1,0 +1,1 @@
+{% element 'Admin/index_content' { 'title': __(resourceType|humanize) } %}
