--- src/Template/Pages/Administration/Applications/index.twig
@@ -1,1 +1,0 @@
-{% element 'Administration/index_content' { 'title': __(resourceType|humanize) } %}
