--- src/Template/Pages/Administration/Config/index.twig
@@ -1,1 +1,0 @@
-{% element 'Administration/index_content' { 'title': __(resourceType|humanize) } %}
